use cluster-wide shared Vault instance

+++ test/e2e/framework/addon/venafi/tpp.go
@@ -46,34 +46,34 @@ type TPPDetails struct {
 	issuerTemplate cmapi.VenafiIssuer
 }
 
-func (v *VenafiTPP) Setup(cfg *config.Config) error {
+func (v *VenafiTPP) Setup(cfg *config.Config, _ ...interface{}) (interface{}, error) {
 	v.config = cfg
 
 	if v.Base == nil {
 		v.Base = &base.Base{}
-		err := v.Base.Setup(cfg)
+		_, err := v.Base.Setup(cfg)
 		if err != nil {
-			return err
+			return nil, err
 		}
 	}
 
 	if v.config.Addons.Venafi.TPP.URL == "" {
-		return errors.NewSkip(fmt.Errorf("Venafi TPP URL must be set"))
+		return nil, errors.NewSkip(fmt.Errorf("Venafi TPP URL must be set"))
 	}
 	if v.config.Addons.Venafi.TPP.Zone == "" {
-		return errors.NewSkip(fmt.Errorf("Venafi TPP Zone must be set"))
+		return nil, errors.NewSkip(fmt.Errorf("Venafi TPP Zone must be set"))
 	}
 
 	if v.config.Addons.Venafi.TPP.AccessToken == "" {
 		if v.config.Addons.Venafi.TPP.Username == "" {
-			return errors.NewSkip(fmt.Errorf("Venafi TPP requires either an access-token or username-password to be set: missing username"))
+			return nil, errors.NewSkip(fmt.Errorf("Venafi TPP requires either an access-token or username-password to be set: missing username"))
 		}
 		if v.config.Addons.Venafi.TPP.Password == "" {
-			return errors.NewSkip(fmt.Errorf("Venafi TPP requires either an access-token or username-password to be set: missing password"))
+			return nil, errors.NewSkip(fmt.Errorf("Venafi TPP requires either an access-token or username-password to be set: missing password"))
 		}
 	}
 
-	return nil
+	return nil, nil
 }
 
 func (v *VenafiTPP) Provision() error {
@@ -112,11 +112,15 @@ func (v *VenafiTPP) Details() *TPPDetails {
 }
 
 func (v *VenafiTPP) Deprovision() error {
+	if v.createdSecret == nil {
+		return nil
+	}
+
 	return v.Base.Details().KubeClient.CoreV1().Secrets(v.createdSecret.Namespace).Delete(context.TODO(), v.createdSecret.Name, metav1.DeleteOptions{})
 }
 
 func (v *VenafiTPP) SupportsGlobal() bool {
-	return true
+	return false
 }
 
 func (t *TPPDetails) BuildIssuer() *cmapi.Issuer {
